Split each build configuration into a separate CI job.  (#546)
Expose the main CMake config knobs as environment variables to tailor build configurations & jobs.

Replace macOS Universal Binary builds with Apple Silicon builds while retaining x86_64 builds.

Add new configs/jobs for testing OpenCL and SSE combos.

Cleanup CI files.

+++ ci_scripts/install_macos.sh
@@ -1,24 +1,30 @@
-#!/bin/sh
+#! /usr/bin/env bash
 # Copyright 2015-2020 The Khronos Group Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-# exit if any command fails
+# Install software in CI environment necessary to build on macOS.
+
+# Exit if any command fails.
 set -e
 
-brew update > /dev/null
-brew install git-lfs
+# Set parameters from command-line arguments, if any.
+for i in $@; do
+  eval $i
+done
+
+FEATURE_LOADTESTS=${FEATURE_LOADTESTS:-ON}
+
 git lfs install
 git lfs version
-brew install doxygen
-brew install sdl2
-brew link sdl2
 gem install xcpretty
 
-# Current directory is .../build/{KhronosGroup,msc-}/KTX-Software. cd to 'build'.
-pushd ../..
-wget -O vulkansdk-macos-$VULKAN_SDK_VER.dmg https://sdk.lunarg.com/sdk/download/$VULKAN_SDK_VER/mac/vulkansdk-macos-$VULKAN_SDK_VER.dmg?Human=true
-hdiutil attach vulkansdk-macos-$VULKAN_SDK_VER.dmg
-sudo /Volumes/vulkansdk-macos-$VULKAN_SDK_VER/InstallVulkan.app/Contents/macOS/InstallVulkan --root "$VULKAN_INSTALL_DIR" --accept-licenses --default-answer --confirm-command install
-hdiutil detach /Volumes/vulkansdk-macos-$VULKAN_SDK_VER
-rm vulkansdk-macos-$VULKAN_SDK_VER.dmg
-popd
+if [ "$FEATURE_LOADTESTS" = "ON" ]; then
+  # Current dir. is .../build/{KhronosGroup,msc-}/KTX-Software. cd to 'build'.
+  pushd ../..
+  wget -O vulkansdk-macos-$VULKAN_SDK_VER.dmg https://sdk.lunarg.com/sdk/download/$VULKAN_SDK_VER/mac/vulkansdk-macos-$VULKAN_SDK_VER.dmg?Human=true
+  hdiutil attach vulkansdk-macos-$VULKAN_SDK_VER.dmg
+  sudo /Volumes/vulkansdk-macos-$VULKAN_SDK_VER/InstallVulkan.app/Contents/macOS/InstallVulkan --root "$VULKAN_INSTALL_DIR" --accept-licenses --default-answer --confirm-command install
+  hdiutil detach /Volumes/vulkansdk-macos-$VULKAN_SDK_VER
+  rm vulkansdk-macos-$VULKAN_SDK_VER.dmg
+  popd
+fi
